requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0

# LLM
anthropic>=0.18.0

# RAG (Milestone 2)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data & Analysis (Milestone 3)
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Code Execution (Milestone 3)
RestrictedPython>=6.0

# Web Scraping (Corpus Expansion)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.0.0